capitalize string function Test Driven Development
Start by writing a test for a capitalize(string) function. My test should make sure that this function takes a string as an argument and returns that string with the first character capitalized.
I made tests green by implementing the capitalize(string) function. Considering the minimum amount of code are necessary to pass this test and write it.